refactor(NcAppContentDetails): migrate component to Typescript
