+++ src/components/NcAppContentDetails/NcAppContentDetails.vue
@@ -3,15 +3,20 @@
   - SPDX-License-Identifier: AGPL-3.0-or-later
 -->
 
+<script setup lang="ts">
+import type { Slot } from 'vue'
+
+defineSlots<{
+	/**
+	 * Provide content to the details view.
+	 * Passing content to this slot is required!
+	 */
+	default?: Slot
+}>()
+</script>
+
 <template>
 	<div class="app-content-details">
-		<!-- @slot Provide content to the details view -->
 		<slot />
 	</div>
 </template>
-
-<script>
-export default {
-	name: 'NcAppContentDetails',
-}
-</script>
