+++ src/components/NcAppContentDetails/index.ts
