+++ src/components/index.ts
@@ -16,7 +16,7 @@ export { default as NcActionSeparator } from './NcActionSeparator/index.js'
 export { default as NcActionText } from './NcActionText/index.js'
 export { default as NcActionTextEditable } from './NcActionTextEditable/index.js'
 export { default as NcAppContent } from './NcAppContent/index.js'
-export { default as NcAppContentDetails } from './NcAppContentDetails/index.js'
+export { default as NcAppContentDetails } from './NcAppContentDetails/index.ts'
 export { default as NcAppContentList } from './NcAppContentList/index.js'
 export { default as NcAppNavigation } from './NcAppNavigation/index.js'
 export { default as NcAppNavigationCaption } from './NcAppNavigationCaption/index.js'
